added Dao/Repository tests & EmployerServiceImplTest

+++ src/main/java/org/example/clientbank/employer/service/EmployerServiceImpl.java
@@ -5,6 +5,7 @@ import org.example.clientbank.employer.Employer;
 import org.example.clientbank.employer.api.dto.RequestEmployerDto;
 import org.example.clientbank.employer.api.dto.RequestPatchEmployerDto;
 import org.example.clientbank.employer.db.EmployerRepository;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
@@ -41,6 +42,9 @@ public class EmployerServiceImpl implements EmployerService {
 
     @Override
     public void deleteById(Long id) {
+        if (!employerRepository.existsById(id)) {
+            throw new EmptyResultDataAccessException("Customer not found with id: " + id, 1);
+        }
         employerRepository.deleteById(id);
     }
